@@ -11,48 +11,48 @@ class UserForm
     /**
      * @var string $first_name
      */
-    private $first_name;
+    private string $first_name;
     /**
      * @var string $last_name
      */
-    private $last_name;
+    private string $last_name;
     /**
      * @var string $password
      */
-    private $password;
+    private string $password;
     /**
      * @var string $passwordConfirm
      */
-    private $passwordConfirm;
+    private string $passwordConfirm;
     /**
      * @var string $email
      * @Assert\Email()
      */
-    private $email;
+    private string $email;
     /**
      * @var string $street
      */
-    private $street;
+    private string $street;
     /**
      * @var string $number
      */
-    private $number;
+    private string $number;
     /**
      * @var string $postal_code
      */
-    private $postal_code;
+    private string $postal_code;
     /**
      * @var string $city
      */
-    private $city;
+    private string $city;
     /**
      * @var string $country
      */
-    private $country;
+    private string $country;
     /**
      * @var string $phone
      */
-    private $phone;
+    private string $phone;
 
     /**
      * @return string
@@ -251,6 +251,4 @@ class UserForm
         $this->phone = $phone;
         return $this;
     }
-
-
 }
