@@ -5,6 +5,7 @@ namespace App\Entity;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
+use Doctrine\ORM\Mapping\JoinColumn;
 use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
@@ -285,8 +286,6 @@ class User extends BaseEntity implements UserInterface
         $this->country = $country;
         return $this;
     }
-
-
 
     /**
      * @see UserInterface
