@@ -2,13 +2,13 @@
 
 namespace App\Form;
 
-use App\Models\Forms\GroupeForm;
+use App\Models\Forms\RoleForm;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\Extension\Core\Type\NumberType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class GroupeType extends AbstractType
+class RoleType extends AbstractType
 {
     /**
      * @param FormBuilderInterface $builder
@@ -17,8 +17,8 @@ class GroupeType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('groupe',TextType::class,[
-                'required' => false
+            ->add('id_role', NumberType::class,[
+                'required'=>true
             ])
         ;
     }
@@ -29,7 +29,7 @@ class GroupeType extends AbstractType
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
-            'data_class' => GroupeForm::class,
+            'data_class' => RoleForm::class,
         ]);
     }
 }
